# api_docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HERMESWA - WhatsApp Multi-Instance API",
    "description": "REST API for WhatsApp Web automation, multi-instance management, and real-time messaging built with Go, Echo framework, and whatsmeow library.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:2121",
      "description": "Local development"
    }
  ],
  "paths": {
    "/api/qr/{instanceId}": {
      "get": {
        "summary": "Login Qr by Instance",
        "deprecated": false,
        "description": "",
        "tags": [
          "Auth Instance"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "instance_id": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "timeout": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "instance_id",
                        "message",
                        "status",
                        "timeout"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "success": true,
                  "message": "QR generation started",
                  "data": {
                    "instance_id": "fab46e665ff031c6dd8e9c95c7d93a62",
                    "message": "QR codes will be sent via WebSocket. Listen to QR_GENERATED event.",
                    "status": "generating",
                    "timeout": "3 minutes"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/status/{instanceId}": {
      "get": {
        "summary": "Check Status by Instance",
        "deprecated": false,
        "description": "",
        "tags": [
          "Auth Instance"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "instanceId": {
                          "type": "string"
                        },
                        "isConnected": {
                          "type": "boolean"
                        },
                        "jid": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "instanceId",
                        "isConnected",
                        "jid"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "success": true,
                  "message": "Status retrieved",
                  "data": {
                    "instanceId": "51029a1911728ac18e0dee34e0ceb97d",
                    "isConnected": true,
                    "jid": "6285148107612:94@s.whatsapp.net"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/instances": {
      "get": {
        "summary": "Check All Instance Status",
        "deprecated": false,
        "description": "",
        "tags": [
          "Auth Instance"
        ],
        "parameters": [
          {
            "name": "all",
            "in": "query",
            "description": "",
            "required": true,
            "example": "true",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/logout/{instanceId}": {
      "post": {
        "summary": "Logout by Instance",
        "deprecated": false,
        "description": "",
        "tags": [
          "Auth Instance"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "instanceId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "instanceId"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "success": true,
                  "message": "Logged out successfully",
                  "data": {
                    "instanceId": "51029a1911728ac18e0dee34e0ceb97d"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/login": {
      "post": {
        "summary": "Create Instance",
        "deprecated": false,
        "description": "",
        "tags": [
          "Auth Instance"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "circle": {
                    "type": "string"
                  }
                },
                "required": [
                  "circle"
                ]
              },
              "example": {
                "circle": "satu"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "instanceId": {
                          "type": "string"
                        },
                        "nextStep": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "instanceId",
                        "nextStep",
                        "status"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "success": true,
                  "message": "Instance created, QR code required",
                  "data": {
                    "instanceId": "fab46e665ff031c6dd8e9c95c7d93a62",
                    "nextStep": "Call GET /qr/:instanceId to get QR code",
                    "status": "qr_required"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/instances/{instanceId}": {
      "delete": {
        "summary": "Delete Instance",
        "deprecated": false,
        "description": "",
        "tags": [
          "Auth Instance"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "instanceId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "instanceId"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "success": true,
                  "message": "Instance deleted successfully",
                  "data": {
                    "instanceId": "fab46e665ff031c6dd8e9c95c7d93a62"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "summary": "Update Instance Fields",
        "deprecated": false,
        "description": "",
        "tags": [
          "Auth Instance"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "used": {
                    "type": "boolean"
                  },
                  "description": {
                    "type": "string"
                  },
                  "circle": {
                    "type": "string"
                  }
                },
                "required": [
                  "used",
                  "description",
                  "circle"
                ]
              },
              "example": {
                "used": true,
                "description": "Backup instance - ready to use",
                "circle": "six"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "instanceId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "instanceId"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "success": true,
                  "message": "Instance updated successfully",
                  "data": {
                    "instanceId": "89d524a43f5782c47e6889d885ae848f"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/qr-cancel/{instanceId}": {
      "delete": {
        "summary": "Cancel Qr Code Instance",
        "deprecated": false,
        "description": "",
        "tags": [
          "Auth Instance"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "instance_id": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "instance_id",
                        "status"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "success": true,
                  "message": "QR generation cancelled successfully",
                  "data": {
                    "instance_id": "fab46e665ff031c6dd8e9c95c7d93a62",
                    "status": "cancelled"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/groups/by-number/{phoneNumber}": {
      "get": {
        "summary": "Get Group",
        "deprecated": false,
        "description": "",
        "tags": [
          "Message/Group/by Phone Number"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/send-group/by-number/{phoneNumber}/media-url": {
      "post": {
        "summary": "Send Message Media URL",
        "deprecated": false,
        "description": "",
        "tags": [
          "Message/Group/by Phone Number"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "groupJid": {
                    "type": "string"
                  },
                  "mediaUrl": {
                    "type": "string"
                  },
                  "caption": {
                    "type": "string"
                  }
                },
                "required": [
                  "groupJid",
                  "mediaUrl",
                  "caption"
                ]
              },
              "example": {
                "groupJid": "120363420893482288@g.us",
                "mediaUrl": "http://192.168.0.31/main/bridging_open/load_dokumen/rss_116_030101202500025360001_1763021707.pdf?token=0283fe31407c3aed25bd12b1e4069611e6dfc6f4",
                "caption": "Check this out!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/by-number/{phoneNumber}": {
      "post": {
        "summary": "Send Message Text",
        "deprecated": false,
        "description": "",
        "tags": [
          "Message/Single/by Phone Number"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "to": {
                    "type": "string"
                  },
                  "message": {
                    "type": "string"
                  }
                },
                "required": [
                  "to",
                  "message"
                ]
              },
              "example": {
                "to": "6285172462751",
                "message": "Hello from HERMESWA!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}